Debug
The assignment 4 is finished. Please note that it didn't parse all data from JSON as it's required in Assignment 6. Instead it parsed a txt file with each line a path to target sprite. The Engine will instantiate all Entities from the file, while still have the player controller independently generated in game.